soong: do not use compression upon creating zip(jar) files
A lot of zip files are created during builds for intermediate purposes
and not using compression significantly shortens build times.

Making these uncompressed allows the host OS to utilize page-cache too.

Note that this change makes the zip files under out/dist/ to also be
uncompressed files, but it won't matter much as the ota zip file's contents
are compressed on its own.

Change-Id: I6e0a96563468e84fce148e63e4658922542b6c26
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>
Signed-off-by: wHo-EM-i <ehteshammalik18998@gmail.com>